Update Makefile to link libraries
'make run' should now still work flawlessly on Mac and Linux but link
the binary with a bunch of libs required to work raylib should now still
work flawlessly on Mac and Linux but link
the binary with a bunch of libs required to work raylib.
I haven't tested this on Mac, but it works on Linux.